some changes i suppose

+++ src/style.rs
@@ -1,49 +1,26 @@
-//use std::collections::HashMap;
-
-//use super::element::Tag;
-
+#[derive(Debug, Clone, Copy)]
 pub struct Pad {
-    pub top: i32,
-    pub low: i32,
-    pub right: i32,
-    pub left: i32,
+    pub top: f32,
+    pub low: f32,
+    pub right: f32,
+    pub left: f32,
 }
 
+#[derive(Debug, Clone, Copy)]
 pub enum Size {
     Window,
-    Man(i32),
+    Man(f32),
     Child,
 }
 
+#[derive(Debug, Clone, Copy)]
 pub enum Gravity {
     Horizontal,
     Vertical,
 }
 
+#[derive(Debug, Clone, Copy)]
 pub struct Style {
     pad: Pad,
     color: [u8; 4],
 }
-
-// pub enum Style {
-//     Box {
-//         dim: (Size, Size),
-//         ceil: (Size, Size),
-//         pad: Pad,
-//         gravity: Gravity,
-//         color: [u8; 4],
-//     },
-//     Text {
-//         font: (&'static str, f32),
-//         color: [u8; 4],
-//         pad: Pad,
-//         spacing: f32,
-//     },
-//     Img {
-//         color: [u8; 4],
-//         pad: Pad,
-//         scale: f32,
-//     },
-// }
-
-//pub type Sheet = HashMap<Tag, Style>;
